release-1.0.0: created limiter logic

+++ src/main/java/com/efedorchenko/timely/limiter/TokenBucket.java
@@ -11,7 +11,7 @@ public interface TokenBucket {
     /**
      * Возвращает текущее количество доступных токенов
      */
-    double getAvailableTokens();
+    int getAvailableTokens();
 
     /**
      * Возвращает максимальную емкость бакета
